Potential fix for code scanning alert no. 2: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using the larger type, and the result will not overflow.

 - Cast one of the operands (items or size) to size_t before the multiplication.
 - This change should be made in the default_bzalloc function in the file bzip-1.0.8/bzlib.c.

+++ bzip-1.0.8/bzlib.c
@@ -101,7 +101,7 @@ int bz_config_ok ( void )
 static
 void* default_bzalloc ( void* opaque, Int32 items, Int32 size )
 {
-   void* v = malloc ( items * size );
+   void* v = malloc ( (size_t)items * size );
    return v;
 }
